requirements.txt
streamlit
numpy
pandas

matplotlib.pyplot
sklearn.preprocessing
sklearn.model_selection
sklearn.linear_model
sklearn.metrics
